FP/requirements.txt
fastapi[standard]
uvicorn[standard]
pytest
pydantic
pytest-asyncio
httpx
scikit-learn
numpy
mlflow
psycopg2-binary
alembic
asyncpg
pgmigrate
aiokafka
redis[asyncio]
python-dotenv
prometheus-client
prometheus-fastapi-instrumentator
sentry-sdk[fastapi]
pyjwt
testcontainers[postgres,redis,kafka]
celery[redis]>=5.3.0